+++ src/main/java/com/checkmarx/configprovider/dto/ResourceType.java
@@ -1,0 +1,9 @@
+package com.checkmarx.configprovider.dto;
+
+public enum ResourceType {
+    YML,
+    JSON,
+    PROPERTIES,
+    ENV_VARIABLES,
+    COMBINED
+}
