+++ src/main/java/com/checkmarx/configprovider/dto/SourceProviderType.java
@@ -1,0 +1,5 @@
+package com.checkmarx.configprovider.dto;
+
+public enum SourceProviderType {
+    GITHUB
+}
