+++ src/main/java/com/checkmarx/configprovider/dto/interfaces/ConfigProperties.java
@@ -1,0 +1,4 @@
+package com.checkmarx.configprovider.dto.interfaces;
+
+public interface ConfigProperties {
+}
