+++ src/main/java/com/checkmarx/configprovider/dto/interfaces/ConfigResource.java
@@ -1,0 +1,7 @@
+package com.checkmarx.configprovider.dto.interfaces;
+
+public interface ConfigResource {
+
+    String getName();
+    
+}
