+++ src/main/java/com/checkmarx/configprovider/exceptions/ConfigProviderException.java
@@ -1,0 +1,11 @@
+package com.checkmarx.configprovider.exceptions;
+
+public class ConfigProviderException extends RuntimeException {
+    public ConfigProviderException(String message, Throwable e) {
+        super(message, e);
+    }
+
+    public ConfigProviderException(String message) {
+        super(message);
+    }
+}
