+++ src/main/java/com/checkmarx/configprovider/interfaces/SourceControlClient.java
@@ -1,0 +1,13 @@
+package com.checkmarx.configprovider.interfaces;
+
+import com.checkmarx.configprovider.dto.RepoDto;
+
+import java.util.List;
+
+public interface SourceControlClient {
+    
+
+    String downloadFileContent(String path, String filename, RepoDto repo);
+
+    List<String> getDirectoryFilenames(RepoDto repoResource, String path);
+}
