+++ src/main/java/com/checkmarx/configprovider/resource/Parser.java
@@ -1,0 +1,13 @@
+package com.checkmarx.configprovider.resource;
+
+import com.checkmarx.configprovider.dto.interfaces.ConfigResource;
+import com.typesafe.config.Config;
+
+import javax.naming.ConfigurationException;
+
+public class Parser {
+    
+    public static Config parse(ConfigResource resource) throws ConfigurationException {
+        return ((ParsableResource)resource).loadConfig();
+    }
+}
