--- src/main/java/com/cx/configprovider/dto/ResourceType.java
@@ -1,9 +1,0 @@
-package com.cx.configprovider.dto;
-
-public enum ResourceType {
-    YML,
-    JSON,
-    PROPERTIES,
-    ENV_VARIABLES,
-    COMBINED
-}
