--- src/main/java/com/cx/configprovider/dto/SourceProviderType.java
@@ -1,5 +1,0 @@
-package com.cx.configprovider.dto;
-
-public enum SourceProviderType {
-    GITHUB
-}
