--- src/main/java/com/cx/configprovider/dto/interfaces/ConfigProperties.java
@@ -1,4 +1,0 @@
-package com.cx.configprovider.dto.interfaces;
-
-public interface ConfigProperties {
-}
