--- src/main/java/com/cx/configprovider/dto/interfaces/ConfigResource.java
@@ -1,7 +1,0 @@
-package com.cx.configprovider.dto.interfaces;
-
-public interface ConfigResource {
-
-    String getName();
-    
-}
