--- src/main/java/com/cx/configprovider/interfaces/SourceControlClient.java
@@ -1,13 +1,0 @@
-package com.cx.configprovider.interfaces;
-
-import com.cx.configprovider.dto.RepoDto;
-
-import java.util.List;
-
-public interface SourceControlClient {
-    
-
-    String downloadFileContent(String path, String filename, RepoDto repo);
-
-    List<String> getDirectoryFilenames(RepoDto repoResource, String path);
-}
