--- src/main/java/com/cx/configprovider/resource/Parser.java
@@ -1,13 +1,0 @@
-package com.cx.configprovider.resource;
-
-import com.cx.configprovider.dto.interfaces.ConfigResource;
-import com.typesafe.config.Config;
-
-import javax.naming.ConfigurationException;
-
-public class Parser {
-    
-    public static Config parse(ConfigResource resource) throws ConfigurationException {
-        return ((ParsableResource)resource).loadConfig();
-    }
-}
